Fix errors on TS
