+++ js/api.js
@@ -5,16 +5,14 @@ const ServerRoute = {
   SEND_DATA: '/',
 };
 
-//перечисления методов запроса на сервер
 const HttpMethod = {
   GET: 'GET',
   POST: 'POST',
 };
 
-//текст ошибок для отправки и получения данных с сервера
 const ErrorText = {
-  [HttpMethod.GET]: 'Не удалось загрузить данные.', //текст сообщения уточнить по ТЗ
-  [HttpMethod.POST]: 'Не удалось отправить данные.', //текст сообщения уточнить по ТЗ
+  [HttpMethod.GET]: 'Не удалось загрузить данные.',
+  [HttpMethod.POST]: 'Не удалось отправить данные.',
 };
 
 async function request(url, method = HttpMethod.GET, body = null) {
@@ -26,12 +24,10 @@ async function request(url, method = HttpMethod.GET, body = null) {
   return response.json();
 }
 
-//функция получения картинки из сервера
 async function getData() {
   return request(`${SERVER_URL}${ServerRoute.GET_DATA}`);
 }
 
-//функция отправки на сервер картинки
 async function postData(pictureData) {
   return request(
     `${SERVER_URL}${ServerRoute.SEND_DATA}`,
