+++ js/creat-templates.js
@@ -1,21 +1,19 @@
-const templatePicture = document.querySelector('#picture').content.querySelector('.picture'); // Нашли шаблон
+const templatePicture = document.querySelector('#picture').content.querySelector('.picture');
 
-// Функция создания DOM-элементов и заполнения их данными
 const createTemplate = ({url, description, comments, likes, id}) => {
-  const thumbnail = templatePicture.cloneNode(true); //Делаем копию шаблона
+  const thumbnail = templatePicture.cloneNode(true);
   thumbnail.querySelector('.picture__img').src = url;
   thumbnail.querySelector('.picture__img').alt = description;
   thumbnail.querySelector('.picture__likes').textContent = likes;
   thumbnail.querySelector('.picture__comments').textContent = comments.length;
 
-  // Присваеваем id для каждой картинки
   thumbnail.dataset.thumbnailId = id;
 
   return thumbnail;
 };
 
 const createTemplates = (pictures, container) => {
-  const pictureListFragment = document.createDocumentFragment(); // Создаем пустой контейнер
+  const pictureListFragment = document.createDocumentFragment();
   pictures.forEach((picture) => {
     const thumbnail = createTemplate(picture);
