+++ js/filters.js
@@ -1,8 +1,7 @@
-import { createTemplates } from './creat_template.js';
+import { createTemplates } from './creat-templates.js';
 import { debounce } from './util.js';
 import { getRandomNumber } from './util.js';
 
-//константа на кол-во случайных фото
 const MAX_RANDOM_FILTER = 10;
 
 const filterElement = document.querySelector('.img-filters');
@@ -12,14 +11,12 @@ const randomButtonElement = filterFormElement.querySelector('#filter-random');
 const discussedButtonElement = filterFormElement.querySelector('#filter-discussed');
 
 
-//объект с фильтрами
 const FilterEnum = {
   DEFAULT: 'default',
   RANDOM: 'random',
   DISCUSSED: 'discussed'
 };
 
-//функция по ключам фильтров
 const filterHandles = {
   [FilterEnum.DEFAULT]: (data) => data,
   [FilterEnum.RANDOM]: (data) => {
@@ -36,32 +33,27 @@ const filterHandles = {
   [FilterEnum.DISCUSSED]: (data) => [...data].sort((item1, item2) => item2.comments.length - item1.comments.length),
 };
 
-//переменная для отмены лишних перерисовок миниатюр при выборе фильтров
 let currentFilter = FilterEnum.DEFAULT;
 
-const repaint = (evt, filter, data) => {
-  //условие для проверки фильтра для отмены лишних перерисовок картинок
+const getRepaint = (evt, filter, data) => {
+
   if (currentFilter !== filter) {
     const container = document.querySelector('.pictures');
     const filteredData = filterHandles[filter](data);
-    //удаление миниатюр перед отрисовкой новых по фильтрам
     const picrutes = document.querySelectorAll('.picture');
     picrutes.forEach((item) => item.remove());
-    //отрисовываем миниатюры
     createTemplates(filteredData, container);
     currentFilter = filter;
   }
 };
 
-//функция для пропуска откликов
-const debounceRepaint = debounce(repaint);
+const debounceRepaint = debounce(getRepaint);
 
 const activateFilter = () => {
   const currentActiveElement = filterFormElement.querySelector('.img-filters__button--active');
   currentActiveElement.classList.remove('img-filters__button--active');
 } ;
 
-//функция по удалению скрытого тега фильтров
 const initFilter = (data) => {
   filterElement.classList.remove('img-filters--inactive');
   defaultButtonElement.addEventListener('click', (evt) => {
