+++ js/scale.js
@@ -4,12 +4,8 @@ const MAX_SCALE = 100;
 const DEFAULT_SCALE = 100;
 
 const modalElement = document.querySelector('.img-upload');
-const smallerButtonElement = modalElement.querySelector(
-  '.scale__control--smaller',
-);
-const biggerButtonElement = modalElement.querySelector(
-  '.scale__control--bigger',
-);
+const smallerButtonElement = modalElement.querySelector('.scale__control--smaller');
+const biggerButtonElement = modalElement.querySelector('.scale__control--bigger');
 const scaleInputElement = modalElement.querySelector('.scale__control--value');
 const imageElement = modalElement.querySelector('.img-upload__preview img');
