+++ js/util.js
@@ -11,7 +11,6 @@ function showErrorMessage() {
   }, REMOVE_MESSAGE_TIMEOUT);
 }
 
-// Функция пропуска откликов
 function debounce (callback, timeoutDelay = REMOVE_MESSAGE_TIMEOUT) {
   let timeoutId;
 
@@ -22,12 +21,10 @@ function debounce (callback, timeoutDelay = REMOVE_MESSAGE_TIMEOUT) {
   };
 }
 
-// Функция для генерации случайного числа в диапазоне от min до max
 function getRandomNumber(min, max) {
   return Math.floor(Math.random() * (max - min + 1)) + min;
 }
 
-// Функция отслеживания нажатия кнопки Esc
 const isEscKeyDown = (evt) => evt.key === 'Escape';
 
 export { getRandomNumber, showErrorMessage, isEscKeyDown, debounce};
